feat: increase map layers
fix: build warning

Code format - (Clang-format)

+++ source/map/map_region.h
@@ -141,9 +141,9 @@ public:
 	}
 
 	void setVisible(bool overground, bool underground);
-	void setVisible(uint32_t client, bool underground, bool value);
-	bool isVisible(uint32_t client, bool underground);
-	void clearVisible(uint32_t client);
+	void setVisible(uint64_t client, bool underground, bool value);
+	bool isVisible(uint64_t client, bool underground);
+	void clearVisible(uint64_t client);
 
 	void setRequested(bool underground, bool r);
 	bool isVisible(bool underground);
@@ -153,7 +153,7 @@ protected:
 	BaseMap &map;
 	QTreeNode* parent;
 	uint32_t subtreeTileCount;
-	uint32_t visible;
+	uint64_t visible;
 
 	bool isLeaf;
